Updated spec to work with package upgrades When yum does a package upgrade, the new version is installed before the old version is removed. This means that the %preun command is ran after the new version has been installed so the init script was no linked. This change checks to see if we are doing an upgrade or a full uninstall before unlinking.

+++ bin/dmda.spec
@@ -50,17 +50,23 @@ rm -rf %{buildroot}
 
 
 %preun
-unlink /etc/init.d/dmdad
+if [ -L /etc/init.d/dmdad ];then
+  # only unlink if this is an uninstall rather than an upgrade
+  if [ "$1" = "0" ];then
+    unlink /etc/init.d/dmdad
+  fi
+fi
 
 
 %pre
 if [ -L /etc/init.d/dmdad ];then
- unlink /etc/init.d/dmdad
+  unlink /etc/init.d/dmdad
 fi
 
 
 %post
 ln -s /opt/dmda/dmdad /etc/init.d/dmdad
+chmod u+x dmdad
 
 
 %changelog
